refactor: re-arrange packages of Sync-related classes
