+++ detekt/baseline.xml
@@ -21,7 +21,7 @@
     <ID>ChainWrapping:com.wire.kalium.logic.feature.message.GetNotificationsUseCase.kt:228</ID>
     <ID>ChainWrapping:com.wire.kalium.logic.feature.message.GetNotificationsUseCase.kt:229</ID>
     <ID>ChainWrapping:com.wire.kalium.logic.feature.message.GetNotificationsUseCase.kt:230</ID>
-    <ID>ChainWrapping:com.wire.kalium.logic.sync.ConversationEventReceiver.kt:297</ID>
+    <ID>ChainWrapping:com.wire.kalium.logic.sync.receiver.ConversationEventReceiver.kt:297</ID>
     <ID>ClassNaming:Cryptobox.module_@wireapp_cryptobox.kt$T$0</ID>
     <ID>ClassNaming:Size_t.kt$Size_t : IntegerType</ID>
     <ID>ClassNaming:Uint32_t.kt$Uint32_t : IntegerType</ID>
@@ -197,8 +197,8 @@
     <ID>CommentSpacing:com.wire.kalium.logic.feature.user.SearchUserUseCaseTest.kt:81</ID>
     <ID>CommentSpacing:com.wire.kalium.logic.feature.user.SearchUserUseCaseTest.kt:93</ID>
     <ID>CommentSpacing:com.wire.kalium.logic.feature.user.SearchUserUseCaseTest.kt:95</ID>
-    <ID>CommentSpacing:com.wire.kalium.logic.sync.ConversationEventReceiver.kt:216</ID>
-    <ID>CommentSpacing:com.wire.kalium.logic.sync.ConversationEventReceiver.kt:46</ID>
+    <ID>CommentSpacing:com.wire.kalium.logic.sync.receiver.ConversationEventReceiver.kt:216</ID>
+    <ID>CommentSpacing:com.wire.kalium.logic.sync.receiver.ConversationEventReceiver.kt:46</ID>
     <ID>CommentSpacing:com.wire.kalium.logic.sync.SyncManager.kt:59</ID>
     <ID>CommentSpacing:com.wire.kalium.logic.sync.SyncManagerTest.kt:113</ID>
     <ID>CommentSpacing:com.wire.kalium.logic.sync.SyncManagerTest.kt:116</ID>
@@ -768,7 +768,7 @@
     <ID>NewLineAtEndOfFile:DomainUserTypeMapperMock.kt$com.wire.kalium.logic.data.user.type.DomainUserTypeMapperMock.kt</ID>
     <ID>NewLineAtEndOfFile:EventGathererMock.kt$com.wire.kalium.logic.sync.EventGathererMock.kt</ID>
     <ID>NewLineAtEndOfFile:EventInfoStorageMock.kt$com.wire.kalium.persistence.event.EventInfoStorageMock.kt</ID>
-    <ID>NewLineAtEndOfFile:EventProcessorMock.kt$com.wire.kalium.logic.sync.event.EventProcessorMock.kt</ID>
+    <ID>NewLineAtEndOfFile:EventProcessorMock.kt$com.wire.kalium.logic.sync.incremental.EventProcessorMock.kt</ID>
     <ID>NewLineAtEndOfFile:EventRepositoryMock.kt$com.wire.kalium.logic.data.event.EventRepositoryMock.kt</ID>
     <ID>NewLineAtEndOfFile:FeatureConfigApiMock.kt$com.wire.kalium.network.api.featureConfigs.FeatureConfigApiMock.kt</ID>
     <ID>NewLineAtEndOfFile:FeatureConfigEventReceiverMock.kt$com.wire.kalium.logic.sync.FeatureConfigEventReceiverMock.kt</ID>
@@ -981,10 +981,10 @@
     <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.functional.Either.kt:73</ID>
     <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.functional.FlowTest.kt:13</ID>
     <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.sync.DefaultWorker.kt:3</ID>
-    <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.sync.PendingEventsBufferTest.kt:58</ID>
+    <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.sync.incremental.PendingEventsBufferTest.kt:58</ID>
     <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.sync.SyncManagerTest.kt:155</ID>
-    <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.sync.UpdateApiVersionsWorker.kt:5</ID>
-    <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.sync.handler.MessageTextEditHandler.kt:41</ID>
+    <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.sync.periodic.UpdateApiVersionsWorker.kt:5</ID>
+    <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.sync.receiver.message.MessageTextEditHandler.kt:41</ID>
     <ID>NoConsecutiveBlankLines:com.wire.kalium.logic.util.stubs.ServerConfigStubs.kt:82</ID>
     <ID>NoConsecutiveBlankLines:com.wire.kalium.network.ServerMetaDataManager.kt:3</ID>
     <ID>NoConsecutiveBlankLines:com.wire.kalium.network.UnauthenticatedNetworkContainer.kt:28</ID>
@@ -1114,7 +1114,7 @@
     <ID>NoUnusedImports:com.wire.kalium.logic.feature.message.MessageEnvelopeCreatorTest.kt:30</ID>
     <ID>NoUnusedImports:com.wire.kalium.logic.feature.message.MessageSenderTest.kt:8</ID>
     <ID>NoUnusedImports:com.wire.kalium.logic.feature.user.UploadUserAvatarUseCaseTest.kt:5</ID>
-    <ID>NoUnusedImports:com.wire.kalium.logic.sync.ConversationEventReceiver.kt:12</ID>
+    <ID>NoUnusedImports:com.wire.kalium.logic.sync.receiver.ConversationEventReceiver.kt:12</ID>
     <ID>NoUnusedImports:com.wire.kalium.logic.sync.WorkSchedulerImpl.kt:9</ID>
     <ID>NoUnusedImports:com.wire.kalium.logic.test_util.TestKaliumDispatcher.kt:6</ID>
     <ID>NoUnusedImports:com.wire.kalium.network.HttpEngine.kt:6</ID>
