+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/sync/WorkSchedulerImpl.kt
@@ -13,6 +13,8 @@ import com.wire.kalium.logger.KaliumLogger.Companion.ApplicationFlow.SYNC
 import com.wire.kalium.logic.CoreLogic
 import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.kaliumLogger
+import com.wire.kalium.logic.sync.full.SlowSyncWorker
+import com.wire.kalium.logic.sync.periodic.UpdateApiVersionsWorker
 import com.wire.kalium.util.KaliumDispatcher
 import com.wire.kalium.util.KaliumDispatcherImpl
 import kotlinx.coroutines.CoroutineScope
