+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/sync/WorkSchedulerImpl.kt
@@ -2,6 +2,8 @@ package com.wire.kalium.logic.sync
 
 import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.feature.message.MessageSendingScheduler
+import com.wire.kalium.logic.sync.full.SlowSyncScheduler
+import com.wire.kalium.logic.sync.periodic.UpdateApiVersionsScheduler
 
 interface GlobalWorkScheduler : UpdateApiVersionsScheduler
 interface UserSessionWorkScheduler : MessageSendingScheduler, SlowSyncScheduler {
