+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/sync/periodic/UpdateApiVersionsScheduler.kt
@@ -1,4 +1,4 @@
-package com.wire.kalium.logic.sync
+package com.wire.kalium.logic.sync.periodic
 
 /**
  * Responsible for [schedulePeriodicApiVersionUpdate] and [scheduleImmediateApiVersionUpdate].
