+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/sync/receiver/EventReceiver.kt
@@ -1,4 +1,4 @@
-package com.wire.kalium.logic.sync
+package com.wire.kalium.logic.sync.receiver
 
 import com.wire.kalium.logic.data.event.Event
