+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/sync/incremental/PendingEventsBufferTest.kt
@@ -1,6 +1,7 @@
-package com.wire.kalium.logic.sync
+package com.wire.kalium.logic.sync.incremental
 
 import com.wire.kalium.logic.framework.TestEvent
+import com.wire.kalium.logic.sync.incremental.PendingEventsBuffer
 import kotlinx.coroutines.test.runTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
